+++ src/main/java/edu/ohsu/cmp/coach/controller/HomeController.java
@@ -10,7 +10,6 @@ import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -19,10 +18,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.client.HttpServerErrorException;
-import org.springframework.web.context.request.RequestContextHolder;
-import org.springframework.web.context.request.ServletRequestAttributes;
 
-import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 import java.util.ArrayList;
 import java.util.List;
@@ -60,8 +56,6 @@ public class HomeController extends BaseController {
             try {
                 model.addAttribute("patient", workspace.getPatient());
                 model.addAttribute("bpGoal", goalService.getCurrentBPGoal(session.getId()));
-
-                model.addAttribute("medicationsOfInterestName", medicationService.getMedicationsOfInterestName());
 
                 List<CDSHook> list = recommendationService.getOrderedCDSHooks();
                 model.addAttribute("cdshooks", list);
@@ -119,7 +113,7 @@ public class HomeController extends BaseController {
     @PostMapping("medications-list")
     public ResponseEntity<List<MedicationModel>> getMedications(HttpSession session) {
         try {
-            List<MedicationModel> list = medicationService.getMedicationsOfInterest(session.getId());
+            List<MedicationModel> list = medicationService.getAntihypertensiveMedications(session.getId());
 
             return new ResponseEntity<>(new ArrayList<>(list), HttpStatus.OK);
