+++ src/main/resources/application.properties
@@ -21,7 +21,7 @@ spring.main.allow-bean-definition-overriding=true
 
 #logging.level.root=INFO
 launch.url=http://localhost:8082/launch-patient
-cqfruler.cdshooks.endpoint.url=http://localhost:8080/cqf-ruler-r4/cds-services
+cqfruler.cdshooks.endpoint.url=http://localhost:8080/cds-services
 cqfruler.cdshooks.order.csv=plandefinition-HypertensiveEmergency,plandefinition-AdverseEvents,plandefinition-Pharma,plandefinition-NonPharmacologicIntervention
 
 # see: https://www.baeldung.com/spring-data-jpa-multiple-databases
@@ -55,3 +55,5 @@ mychart.askAMedicalQuestion.url=https://mychartweb.ohsu.edu/MyChart/inside.asp?m
 
 fhir.search.count=1000
 fhir.observation.bp-write=true
+
+antihypertensive.medication.valueset.oid.csv=2.16.840.1.113762.1.4.1178.10,2.16.840.1.113762.1.4.1178.72
