+++ src/main/resources/fhirconfig.properties
@@ -56,9 +56,3 @@ protocol.answer.system=http://loinc.org
 protocol.answer.display=Other, Specify:
 protocol.answer.yes=Followed BP Protocol
 protocol.answer.no=Did NOT Follow BP Protocol
-
-############################################
-# Medication
-############################################
-
-medication.valueset.oid=2.16.840.1.113762.1.4.1178.10
