+++ src/main/resources/static/js/ie/chart.js-2.9.3/LICENSE.md
