+++ src/main/resources/templates/medications.mustache
@@ -41,13 +41,13 @@
         <table class="content">
             <tr>
                 <td class="heading">
-                    My {{medicationsOfInterestName}}
+                    My Antihypertensive Medications
                 </td>
             </tr>
             <tr>
                 <td>
-                    <table id=medicationsOfInterestTable" class="data">
-                        <tr id="medicationsOfInterestTableHeader">
+                    <table id=antihypertensiveMedicationsTable" class="data">
+                        <tr id="antihypertensiveMedicationsTableHeader">
                             <th>Name</th>
                             <th>Reason</th>
                             <th>Dose</th>
@@ -55,7 +55,7 @@
                             <th>Issues</th>
                             <th>Priority</th>
                         </tr>
-                        {{#medicationsOfInterest}}
+                        {{#antihypertensiveMedications}}
                             <tr class='medication' data-system="{{system}}" data-code="{{code}}">
                                 <td>{{description}}</td>
                                 <td>{{#reason}}{{reason}}{{/reason}}</td>
@@ -64,7 +64,7 @@
                                 <td>{{#issues}}{{issues}}{{/issues}}</td>
                                 <td>{{#priority}}{{priority}}{{/priority}}</td>
                             </tr>
-                        {{/medicationsOfInterest}}
+                        {{/antihypertensiveMedications}}
                     </table>
                 </td>
             </tr>
